Only delete old posted jobs if there's more than 5

+++ lib/tasks/careers.rake
@@ -2,7 +2,7 @@ namespace :careers do
   desc "purges old job entries"
 
   task update_jobs: :environment do
-    Job.where('post_date < ?', DateTime.now-30.days).destroy_all
+    Job.where('post_date < ?', DateTime.now-30.days).destroy_all if Job.count > 4
   end
 
   task post_to_slack: :environment do
